cloud_storage: move reusable pieces of remote_segment
This moves the following from remote_segment into more accessible
locations:
- cache timeouts are moved into cache_service.h
- download_exception is placed into its own .cc file
