+++ src/v/cloud_storage/CMakeLists.txt
@@ -16,6 +16,7 @@ v_cc_library(
     cache_service.cc
     access_time_tracker.cc
     cache_probe.cc
+    download_exception.cc
     topic_manifest.cc
     partition_manifest.cc
     recursive_directory_walker.cc
