+++ src/v/cloud_storage/download_exception.cc
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2023 Redpanda Data, Inc.
+ *
+ * Licensed as a Redpanda Enterprise file under the Redpanda Community
+ * License (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * https://github.com/redpanda-data/redpanda/blob/master/licenses/rcl.md
+ */
+
+#include "cloud_storage/download_exception.h"
+
+#include "vassert.h"
+
+namespace cloud_storage {
+
+download_exception::download_exception(
+  download_result r, std::filesystem::path p)
+  : result(r)
+  , path(std::move(p)) {
+    vassert(
+      r != download_result::success,
+      "Exception created with successful error code");
+}
+
+const char* download_exception::what() const noexcept {
+    switch (result) {
+    case download_result::failed:
+        return "Failed";
+    case download_result::notfound:
+        return "NotFound";
+    case download_result::timedout:
+        return "TimedOut";
+    case download_result::success:
+        vassert(false, "Successful result can't be used as an error");
+    }
+    __builtin_unreachable();
+}
+
+} // namespace cloud_storage
