+++ src/v/cloud_storage/download_exception.h
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2023 Redpanda Data, Inc.
+ *
+ * Licensed as a Redpanda Enterprise file under the Redpanda Community
+ * License (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * https://github.com/redpanda-data/redpanda/blob/master/licenses/rcl.md
+ */
+#pragma once
+
+#include "cloud_storage/types.h"
+
+namespace cloud_storage {
+
+class download_exception : public std::exception {
+public:
+    explicit download_exception(download_result r, std::filesystem::path p);
+
+    const char* what() const noexcept override;
+
+    const download_result result;
+    std::filesystem::path path;
+};
+
+} // namespace cloud_storage
