+++ src/v/cloud_storage/tests/remote_partition_fuzz_test.cc
@@ -10,6 +10,7 @@
  */
 
 #include "cloud_storage/async_manifest_view.h"
+#include "cloud_storage/download_exception.h"
 #include "cloud_storage/tests/cloud_storage_fixture.h"
 #include "cloud_storage/tests/s3_imposter.h"
 #include "cloud_storage/tests/util.h"
